remove out-of-date fluid APIs and unitest(s), including density_prior_box, anchor_generator, roi_perspective_transform, generate_proposal_labels, generate_mask_labels, generate_proposals, box_clip, retinanet_detection_output, multiclass_nms, locality_aware_nms, matrix_nms, distribute_fpn_proposals, box_decoder_and_assign, collect_fpn_proposals
