+++ python/paddle/fluid/tests/unittests/test_locality_aware_nms_op.py
@@ -422,138 +422,5 @@ class TestLocalityAwareNMSAPI(unittest.TestCase):
         )
 
 
-class TestLocalityAwareNMSError(unittest.TestCase):
-    def test_error(self):
-        boxes = fluid.data(name='bboxes', shape=[None, 81, 8], dtype='float32')
-        scores = fluid.data(name='scores', shape=[None, 1, 81], dtype='float32')
-
-        boxes_int = fluid.data(
-            name='bboxes_int', shape=[None, 81, 8], dtype='int32'
-        )
-        scores_int = fluid.data(
-            name='scores_int', shape=[None, 1, 81], dtype='int32'
-        )
-        boxes_tmp = [1, 2]
-        scores_tmp = [1, 2]
-
-        # type of boxes and scores must be variable
-        self.assertRaises(
-            TypeError,
-            fluid.layers.locality_aware_nms,
-            boxes_tmp,
-            scores,
-            0.5,
-            400,
-            200,
-        )
-        self.assertRaises(
-            TypeError,
-            fluid.layers.locality_aware_nms,
-            boxes,
-            scores_tmp,
-            0.5,
-            400,
-            200,
-        )
-
-        # dtype of boxes and scores must in ['float32', 'float64']
-        self.assertRaises(
-            TypeError,
-            fluid.layers.locality_aware_nms,
-            boxes_int,
-            scores,
-            0.5,
-            400,
-            200,
-        )
-        self.assertRaises(
-            TypeError,
-            fluid.layers.locality_aware_nms,
-            boxes,
-            scores_int,
-            0.5,
-            400,
-            200,
-        )
-
-        score_threshold = int(1)
-        # type of score_threshold must be float
-        self.assertRaises(
-            TypeError,
-            fluid.layers.locality_aware_nms,
-            boxes,
-            scores,
-            score_threshold,
-            400,
-            200,
-        )
-
-        nms_top_k = 400.5
-        # type of num_top_k must be int
-        self.assertRaises(
-            TypeError,
-            fluid.layers.locality_aware_nms,
-            boxes,
-            scores,
-            0.5,
-            nms_top_k,
-            200,
-        )
-
-        keep_top_k = 200.5
-        # type of keep_top_k must be int
-        self.assertRaises(
-            TypeError,
-            fluid.layers.locality_aware_nms,
-            boxes,
-            scores,
-            0.5,
-            400,
-            keep_top_k,
-        )
-
-        nms_threshold = int(0)
-        # type of nms_threshold must be int
-        self.assertRaises(
-            TypeError,
-            fluid.layers.locality_aware_nms,
-            boxes,
-            scores,
-            0.5,
-            400,
-            200,
-            nms_threshold,
-        )
-
-        nms_eta = int(1)
-        # type of nms_eta must be float
-        self.assertRaises(
-            TypeError,
-            fluid.layers.locality_aware_nms,
-            boxes,
-            scores,
-            0.5,
-            400,
-            200,
-            0.5,
-            nms_eta,
-        )
-
-        bg_label = 1.5
-        # type of background_label must be int
-        self.assertRaises(
-            TypeError,
-            fluid.layers.locality_aware_nms,
-            boxes,
-            scores,
-            0.5,
-            400,
-            200,
-            0.5,
-            1.0,
-            bg_label,
-        )
-
-
 if __name__ == '__main__':
     unittest.main()
