+++ python/paddle/fluid/tests/unittests/test_matrix_nms_op.py
@@ -16,8 +16,6 @@ import unittest
 import numpy as np
 import copy
 from op_test import OpTest
-import paddle.fluid as fluid
-from paddle.fluid import Program, program_guard
 import paddle
 
 
@@ -303,124 +301,6 @@ class TestMatrixNMSOpGaussian(TestMatrixNMSOp):
     def set_argument(self):
         self.post_threshold = 0.0
         self.use_gaussian = True
-
-
-class TestMatrixNMSError(unittest.TestCase):
-    def test_errors(self):
-        M = 1200
-        N = 7
-        C = 21
-        BOX_SIZE = 4
-        nms_top_k = 400
-        keep_top_k = 200
-        score_threshold = 0.01
-        post_threshold = 0.0
-
-        boxes_np = np.random.random((M, C, BOX_SIZE)).astype('float32')
-        scores = np.random.random((N * M, C)).astype('float32')
-        scores = np.apply_along_axis(softmax, 1, scores)
-        scores = np.reshape(scores, (N, M, C))
-        scores_np = np.transpose(scores, (0, 2, 1))
-
-        with program_guard(Program(), Program()):
-            boxes_data = fluid.data(
-                name='bboxes', shape=[M, C, BOX_SIZE], dtype='float32'
-            )
-            scores_data = fluid.data(
-                name='scores', shape=[N, C, M], dtype='float32'
-            )
-
-            def test_bboxes_Variable():
-                # the bboxes type must be Variable
-                fluid.layers.matrix_nms(
-                    bboxes=boxes_np,
-                    scores=scores_data,
-                    score_threshold=score_threshold,
-                    post_threshold=post_threshold,
-                    nms_top_k=nms_top_k,
-                    keep_top_k=keep_top_k,
-                )
-                paddle.vision.ops.matrix_nms(
-                    bboxes=boxes_np,
-                    scores=scores_data,
-                    score_threshold=score_threshold,
-                    post_threshold=post_threshold,
-                    nms_top_k=nms_top_k,
-                    keep_top_k=keep_top_k,
-                )
-
-            def test_scores_Variable():
-                # the scores type must be Variable
-                fluid.layers.matrix_nms(
-                    bboxes=boxes_data,
-                    scores=scores_np,
-                    score_threshold=score_threshold,
-                    post_threshold=post_threshold,
-                    nms_top_k=nms_top_k,
-                    keep_top_k=keep_top_k,
-                )
-                paddle.vision.ops.matrix_nms(
-                    bboxes=boxes_data,
-                    scores=scores_np,
-                    score_threshold=score_threshold,
-                    post_threshold=post_threshold,
-                    nms_top_k=nms_top_k,
-                    keep_top_k=keep_top_k,
-                )
-
-            def test_empty():
-                # when all score are lower than threshold
-                try:
-                    fluid.layers.matrix_nms(
-                        bboxes=boxes_data,
-                        scores=scores_data,
-                        score_threshold=score_threshold,
-                        post_threshold=post_threshold,
-                        nms_top_k=nms_top_k,
-                        keep_top_k=keep_top_k,
-                    )
-                except Exception as e:
-                    self.fail(e)
-                try:
-                    paddle.vision.ops.matrix_nms(
-                        bboxes=boxes_data,
-                        scores=scores_data,
-                        score_threshold=score_threshold,
-                        post_threshold=post_threshold,
-                        nms_top_k=nms_top_k,
-                        keep_top_k=keep_top_k,
-                    )
-                except Exception as e:
-                    self.fail(e)
-
-            def test_coverage():
-                # cover correct workflow
-                try:
-                    fluid.layers.matrix_nms(
-                        bboxes=boxes_data,
-                        scores=scores_data,
-                        score_threshold=score_threshold,
-                        post_threshold=post_threshold,
-                        nms_top_k=nms_top_k,
-                        keep_top_k=keep_top_k,
-                    )
-                except Exception as e:
-                    self.fail(e)
-                try:
-                    paddle.vision.ops.matrix_nms(
-                        bboxes=boxes_data,
-                        scores=scores_data,
-                        score_threshold=score_threshold,
-                        post_threshold=post_threshold,
-                        nms_top_k=nms_top_k,
-                        keep_top_k=keep_top_k,
-                    )
-                except Exception as e:
-                    self.fail(e)
-
-            self.assertRaises(TypeError, test_bboxes_Variable)
-            self.assertRaises(TypeError, test_scores_Variable)
-            test_coverage()
 
 
 if __name__ == '__main__':
