+++ python/paddle/fluid/tests/unittests/test_retinanet_detection_output.py
@@ -18,8 +18,6 @@ import math
 from op_test import OpTest
 from test_anchor_generator_op import anchor_generator_in_python
 from test_multiclass_nms_op import nms
-import paddle.fluid as fluid
-from paddle.fluid import Program, program_guard
 import paddle
 
 
@@ -504,132 +502,6 @@ class TestRetinanetDetectionOutOpNo5(TestRetinanetDetectionOutOp1):
         for i in range(num_levels):
             self.layer_h.append(2 ** (num_levels - i))
             self.layer_w.append(2 ** (num_levels - i))
-
-
-class TestRetinanetDetectionOutOpError(unittest.TestCase):
-    def test_errors(self):
-        with program_guard(Program(), Program()):
-            bboxes_low1 = fluid.data(
-                name='bboxes_low1', shape=[1, 44, 4], dtype='float32'
-            )
-            bboxes_high1 = fluid.data(
-                name='bboxes_high1', shape=[1, 11, 4], dtype='float32'
-            )
-            scores_low1 = fluid.data(
-                name='scores_low1', shape=[1, 44, 10], dtype='float32'
-            )
-            scores_high1 = fluid.data(
-                name='scores_high1', shape=[1, 11, 10], dtype='float32'
-            )
-            anchors_low1 = fluid.data(
-                name='anchors_low1', shape=[44, 4], dtype='float32'
-            )
-            anchors_high1 = fluid.data(
-                name='anchors_high1', shape=[11, 4], dtype='float32'
-            )
-            im_info1 = fluid.data(
-                name="im_info1", shape=[1, 3], dtype='float32'
-            )
-
-            # The `bboxes` must be list, each element must be Variable and
-            # its Tensor data type must be one of float32 and float64.
-            def test_bboxes_type():
-                fluid.layers.retinanet_detection_output(
-                    bboxes=bboxes_low1,
-                    scores=[scores_low1, scores_high1],
-                    anchors=[anchors_low1, anchors_high1],
-                    im_info=im_info1,
-                )
-
-            self.assertRaises(TypeError, test_bboxes_type)
-
-            def test_bboxes_tensor_dtype():
-                bboxes_high2 = fluid.data(
-                    name='bboxes_high2', shape=[1, 11, 4], dtype='int32'
-                )
-                fluid.layers.retinanet_detection_output(
-                    bboxes=[bboxes_high2, 5],
-                    scores=[scores_low1, scores_high1],
-                    anchors=[anchors_low1, anchors_high1],
-                    im_info=im_info1,
-                )
-
-            self.assertRaises(TypeError, test_bboxes_tensor_dtype)
-
-            # The `scores` must be list, each element must be Variable and its
-            # Tensor data type must be one of float32 and float64.
-            def test_scores_type():
-                fluid.layers.retinanet_detection_output(
-                    bboxes=[bboxes_low1, bboxes_high1],
-                    scores=scores_low1,
-                    anchors=[anchors_low1, anchors_high1],
-                    im_info=im_info1,
-                )
-
-            self.assertRaises(TypeError, test_scores_type)
-
-            def test_scores_tensor_dtype():
-                scores_high2 = fluid.data(
-                    name='scores_high2', shape=[1, 11, 10], dtype='int32'
-                )
-                fluid.layers.retinanet_detection_output(
-                    bboxes=[bboxes_low1, bboxes_high1],
-                    scores=[scores_high2, 5],
-                    anchors=[anchors_low1, anchors_high1],
-                    im_info=im_info1,
-                )
-
-            self.assertRaises(TypeError, test_scores_tensor_dtype)
-
-            # The `anchors` must be list, each element must be Variable and its
-            # Tensor data type must be one of float32 and float64.
-            def test_anchors_type():
-                fluid.layers.retinanet_detection_output(
-                    bboxes=[bboxes_low1, bboxes_high1],
-                    scores=[scores_low1, scores_high1],
-                    anchors=anchors_low1,
-                    im_info=im_info1,
-                )
-
-            self.assertRaises(TypeError, test_anchors_type)
-
-            def test_anchors_tensor_dtype():
-                anchors_high2 = fluid.data(
-                    name='anchors_high2', shape=[11, 4], dtype='int32'
-                )
-                fluid.layers.retinanet_detection_output(
-                    bboxes=[bboxes_low1, bboxes_high1],
-                    scores=[scores_low1, scores_high1],
-                    anchors=[anchors_high2, 5],
-                    im_info=im_info1,
-                )
-
-            self.assertRaises(TypeError, test_anchors_tensor_dtype)
-
-            # The `im_info` must be Variable and the data type of `im_info`
-            # Tensor must be one of float32 and float64.
-            def test_iminfo_type():
-                fluid.layers.retinanet_detection_output(
-                    bboxes=[bboxes_low1, bboxes_high1],
-                    scores=[scores_low1, scores_high1],
-                    anchors=[anchors_low1, anchors_high1],
-                    im_info=[2, 3, 4],
-                )
-
-            self.assertRaises(TypeError, test_iminfo_type)
-
-            def test_iminfo_tensor_dtype():
-                im_info2 = fluid.data(
-                    name='im_info2', shape=[1, 3], dtype='int32'
-                )
-                fluid.layers.retinanet_detection_output(
-                    bboxes=[bboxes_low1, bboxes_high1],
-                    scores=[scores_low1, scores_high1],
-                    anchors=[anchors_low1, anchors_high1],
-                    im_info=im_info2,
-                )
-
-            self.assertRaises(TypeError, test_iminfo_tensor_dtype)
 
 
 if __name__ == '__main__':
