+++ tools/parallel_UT_rule.py
@@ -706,7 +706,6 @@ SECONDARY_HIGH_PARALLEL_JOB_NEW = [
     'test_uniform_random_bf16_op',
     'test_custom_concat',
     'test_weight_quantization_mobilenetv1',
-    'test_retinanet_detection_output',
     'test_concat_mkldnn_op',
     'test_gaussian_random_mkldnn_op',
     'test_parallel_executor_seresnext_with_reduce_cpu',
@@ -1627,7 +1626,6 @@ SIXTH_PARALLEL_JOB_NEW = [
     'test_trt_tuned_dynamic_shape',
     'test_trt_convert_activation',
     'test_trt_convert_affine_channel',
-    'test_trt_convert_anchor_generator',
     'test_trt_fc_fuse_quant_dequant_pass',
     'test_trt_convert_batch_norm',
     'test_trt_conv_quant_dequant_pass',
@@ -1760,7 +1758,6 @@ CPU_PARALLEL_JOB = [
     'test_rpn_target_assign_op',
     'test_row_conv',
     'test_rnn_memory_helper_op',
-    'test_retinanet_detection_output',
     'test_reshape_transpose_matmul_mkldnn_fuse_pass',
     'test_reshape_bf16_op',
     'test_require_version',
@@ -2523,7 +2520,6 @@ TETRAD_PARALLEL_JOB = [
     'test_where_index',
     'test_variance_layer',
     'test_unsqueeze_op',
-    'test_trt_anchor_generator_op',
     'test_translated_layer',
     'test_tensor_shape',
     'test_slice',
@@ -3298,7 +3294,6 @@ TWO_PARALLEL_JOB = [
     'test_trt_convert_tile',
     'test_trt_convert_nearest_interp_v2',
     'test_trt_convert_layer_norm',
-    'test_trt_convert_anchor_generator',
     'test_trt_convert_clip',
     'test_trt_convert_batch_norm',
     'test_cuda_stream_event',
