+++ tools/static_mode_white_list.py
@@ -444,7 +444,6 @@ STATIC_MODE_TESTING_LIST = [
     'test_reorder_lod_tensor',
     'test_reshape_op',
     'test_reshape_bf16_op',
-    'test_retinanet_detection_output',
     'test_reverse_op',
     'test_rmsprop_op',
     'test_rnn_cell_api',
